frontend/requirements.txt
# Amulet AI Frontend Requirements
# ไลบรารีที่จำเป็นสำหรับ Frontend Dashboard

# Core Streamlit
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.15.0

# API Communication  
requests>=2.31.0

# Image Processing
Pillow>=10.0.0

# Date and Time
python-dateutil>=2.8.2

# Optional: Additional UI Components
streamlit-option-menu>=0.3.6
streamlit-aggrid>=0.3.4